Corrigiendo bugs y malas practicas

+++ js/carrito.js
@@ -1,37 +1,35 @@
 /*
- * @ version: v2.0.3
+ * @ version: v3.0.0
  * @ autor: Mirko Pescio
- * @ fecha: 27/12/2021
- * @ Descripción: 2da Preentrega del proyecto final
+ * @ fecha: 09/02/2022
+ * @ Descripción: reentrega del proyecto final
 */
 
 class Carrito {
-    // Primer método
 
-    // Añadir el producto al carrito de compras
+// Añadir el producto al carrito de compras
     comprarProducto(e) {
         e.preventDefault();
         if (e.target.classList.contains("agregar-carrito")) { // Una vez clickeado el botón para agregar al carrito se va a definir lo siguiente
             const producto = e.target.parentElement.parentElement.parentElement;
             this.leerDatosProducto(producto);
-            //console.log(producto);
         }
     }
 
+// Asigno un objeto a cada producto seleccionado
     leerDatosProducto(producto) {
-        const infoProducto = { // Asigno un objeto a cada producto seleccionado
+        const infoProducto = {
             imagen: producto.querySelector("img").src,
             titulo: producto.querySelector("h5").textContent,
             precio: producto.querySelector(".precioProducto").textContent,
             id: producto.querySelector("button").getAttribute("data-id"),
             cantidad: 1
         }
-        // Defino un algoritmo para el carrito de compras:
-        // Este algoritmo que encarga de que no se duplique un mismo producto
-        // en el carrito de compras
-        // y mostrar por pantalla un aviso de que el producto fué agregado anteriormente
-        let productosLS;
-        productosLS = this.obtenerProductosLocalStorage();
+// Defino un algoritmo para el carrito de compras:
+// Este algoritmo que encarga de que no se duplique un mismo producto
+// en el carrito de compras
+// y mostrar por pantalla un aviso de que el producto fué agregado anteriormente
+        let productosLS = this.obtenerProductosLocalStorage();
         productosLS.forEach(function(productoLS){
             if(productoLS.id === infoProducto.id) {
                 productosLS = productoLS.id;
@@ -41,12 +39,13 @@ class Carrito {
             swal({
                 title: 'Oops...',
                 icon: "error",
-                text: 'El producto seleccionado ya fué agregado',
-                timer: 2000, // 2 segundos
-                button: false
+                text: 'El producto seleccionado ya fué agregado, modifique después al procesar la compra',
+                timer: 8000, // 8 segundos
+                button: true
             });
+            infoProducto.cantidad += 1; 
 
-            // clase 13: animaciones___ejercicio_complementario
+        // clase 13: animaciones___ejercicio_complementario
             $(`.card[data-id = "${infoProducto.id}"]`).append(`<p class="agregado" style="display: none">El producto ya fué agregado al carrito de compras</p>`);
             $(".agregado")
                 .fadeIn(3000)
@@ -57,10 +56,11 @@ class Carrito {
                 "font-size": "1.25rem",
                 "border": "2px solid #ccc"
             });
-            setTimeout(function() { // Es para eliminar el parrafo ".agregado" después de la animación
+            // y elimino el parrafo ".agregado" después de la animación
+            setTimeout(function() {
                 $(".agregado")
                     .remove()
-            }, 6500);
+            }, 5000);
         }
         else {
             this.insertarCarrito(infoProducto);
@@ -74,7 +74,8 @@ class Carrito {
         }
     }
 
-    insertarCarrito(producto) { // Aplico en el DOM la estructura de cada producto en el carrito de compras
+// Aplico en el DOM la estructura de cada producto en el carrito de compras
+    insertarCarrito(producto) {
         const row = document.createElement("tr");
         row.innerHTML = `
             <td>
@@ -86,13 +87,13 @@ class Carrito {
                 <a href="#" class="fas fa-times-circle borrar-producto" data-id="${producto.id}"></a>
             </td>`;
         listaProductos.appendChild(row);
-        // Asigno cada producto asignado al localStorage (ver funciones desde la línea 105)
+// Asigno cada producto asignado al localStorage
         this.guardarProductosLocalStorage(producto);
-        // console.log(row);   Los datos se están agregando correctamente
     }
 
-    eliminarProducto(e) { // Para eliminar productos individualmente
-        // Al clickear el icono remove de dicho producto, lo borra del carrito y del localStorage
+// Para eliminar productos individualmente
+    eliminarProducto(e) {
+    // Al clickear el icono remove de dicho producto, lo borra del carrito y del localStorage
         e.preventDefault();
         let producto, productoID;
         if (e.target.classList.contains("borrar-producto")) {
@@ -101,10 +102,12 @@ class Carrito {
             productoID = producto.querySelector("a").getAttribute("data-id");
         }
         this.eliminarProductosLocalStorage(productoID);
-        this.calcularTotal(); // Es para que el monto total se actualice si se borra algún producto
+    // Para que el monto total se actualice si se borra algún producto
+        this.calcularTotal();
     }
 
-    vaciarCarrito(e) { // Para vaciar todo el carrito de compras y el local Storage con un sólo botón
+    // Para vaciar todo el carrito de compras y el local Storage con un sólo botón
+    vaciarCarrito(e) { 
         e.preventDefault();
         while(listaProductos.firstChild) {
             listaProductos.removeChild(listaProductos.firstChild);
@@ -113,14 +116,18 @@ class Carrito {
         return false;
     }
 
-    guardarProductosLocalStorage(producto) { // Una vez creado el array de productos del localStorage, pusheamos los productos que el usuario asigne al carrito de compras
+    // Una vez creado el array de productos del localStorage,
+    // pusheamos los productos que el usuario asigne al carrito de compras
+    guardarProductosLocalStorage(producto) {
         let productos;
         productos = this.obtenerProductosLocalStorage();
         productos.push(producto);
         localStorage.setItem("productos", JSON.stringify(productos));
     }
 
-    obtenerProductosLocalStorage() { // Inicializamos un array de objetos en el localStorage en el caso de que dicho array esté vacío
+    // Inicializamos un array de objetos en el localStorage en el caso 
+    // de que dicho array esté vacío
+    obtenerProductosLocalStorage() {
         let productoLS;
         if(localStorage.getItem("productos") === null) {
             productoLS = [];
@@ -131,21 +138,25 @@ class Carrito {
         return productoLS;
     }
 
-    eliminarProductosLocalStorage(productoID) { // Si eliminamos un producto del carrito de compras, también tenemos que eliminarlo del localStorage
+    // Si eliminamos un producto del carrito de compras, 
+    // también tenemos que eliminarlo del localStorage
+    eliminarProductosLocalStorage(productoID) {
         let productosLS;
         productosLS = this.obtenerProductosLocalStorage();
         productosLS.forEach(function(productoLS, index){
             if(productoLS.id === productoID) {
-                productosLS.splice(index, 1); // Borra un elemento en la posición Index
+                productosLS.splice(index, 1);
+                // Borra un elemento en la posición index
             }
         });
         localStorage.setItem("productos", JSON.stringify(productosLS));
-        // Esta función la llamamos en la función eliminarProducto(e)___Línea 92
+        // Esta función la llamamos en la función eliminarProducto(e)
     }
 
-    leerLocalStorage() { // Esta función es para cuando apenas se cargue la página,
-                         // Los productos que estén en el localStorage, sean automaticamente
-                         // Asignados al carrito de compras
+    // Esta función es para cuando apenas se cargue la página,
+    // Los productos que estén en el localStorage, sean automaticamente
+    // Asignados al carrito de compras
+    leerLocalStorage() {
         let productosLS;
         productosLS = this.obtenerProductosLocalStorage();
         productosLS.forEach(function(producto){
@@ -163,8 +174,8 @@ class Carrito {
         });
     }
 
-    leerLocalStorageCompra() { // Interfaz de los productos en la sección de
-                               // compra (compra.html)
+    // Interfaz de los productos en la sección de compra (compra.html)
+    leerLocalStorageCompra() {
         let productosLS;
         productosLS = this.obtenerProductosLocalStorage();
         productosLS.forEach(function(producto){
@@ -185,22 +196,38 @@ class Carrito {
             listaCompra.appendChild(row);
         });
     }
-
     vaciarLocalStorage() {
-        localStorage.clear();
+        if (this.obtenerProductosLocalStorage().length !== 0){
+            localStorage.clear();
+            swal({
+                title: '¡SE VACIÓ EL CARRITO!',
+                icon: "info",
+                timer: 2000,
+                button: false
+            });
+        }
+        else{
+            swal({
+                title: '¡El CARRITO YA ESTÁ VACÍO!',
+                icon: "error",
+                timer: 2000,
+                button: false
+            });
+        }
     }
 
-    procesarPedidoIndex(e){ // Si hay algún producto agregado en el carrito de compras
-                            // y por lo tanto también en el localStorage, al clickear
-                            // el botón de procesar compra del carrito, esto nos
-                            // redirige a la sección para procesar la compra (compra.html)
+    // Si hay algún producto agregado en el carrito de compras
+    // y por lo tanto también en el localStorage, al clickear
+    // el botón de procesar compra del carrito, esto nos
+    // redirige a la sección para procesar la compra (compra.html)
+    procesarPedidoIndex(e){
         e.preventDefault();
         if (this.obtenerProductosLocalStorage().length === 0){
             swal({
                 title: '¡ACCESO DENEGADO!',
                 icon: "error",
                 text: 'Tu carrito de compras está vacío',
-                timer: 2000, // 2 segundos
+                timer: 2000,
                 button: false
             });
         }
@@ -209,17 +236,18 @@ class Carrito {
         }
     }
 
-    procesarPedidoSecciones(e){ // Es lo mismo que la función procesarPedidoIndex (e)
-                                // Nada mas que esta función se aplica para las otras
-                                // secciones html por una cuestión de las rutas
-                                // La finalidad es ir al archivo compra.html
+    // Es lo mismo que la función procesarPedidoIndex (e)
+    // Nada mas que esta función se aplica para las otras
+    // secciones html por una cuestión de las rutas
+    // La finalidad es ir al archivo compra.html
+    procesarPedidoSecciones(e){
         e.preventDefault();
         if (this.obtenerProductosLocalStorage().length === 0){
             swal({
                 title: '¡ACCESO DENEGADO!',
                 icon: "error",
                 text: 'Tu carrito de compras está vacío',
-                timer: 2000, // 2 segundos
+                timer: 2000,
                 button: false
             });
         }
@@ -228,22 +256,24 @@ class Carrito {
         }
     }
 
-    calcularTotal() { // Calculo del total de cada producto según la cantidad establecida
+    // Calculo del total de cada producto según la cantidad establecida
+    calcularTotal() {
         let productoLS;
         let total = 0;
         productoLS = this.obtenerProductosLocalStorage();
-        for (let index = 0; index < productoLS.length; index++) {
-            let element = Number(productoLS[index].precio * productoLS[index].cantidad);
-            total += element;
+        for (let i = 0; i < productoLS.length; i++) {
+            let n = Number(productoLS[i].precio * productoLS[i].cantidad);
+            total += n;
         }
         document.getElementById("total").innerHTML = "$ " + total.toFixed(2) // Para mostrar 2 decimales
     }
 
-    obtenerEvento(e) { // Esta función es para la actualización de cada subtotal
-                       // de los productos, como para la actualización del valor
-                       // total de la compra.
-                       // Dichos cambios se generan al modificar la cantidad
-                       // seleccionada para cada producto
+    // Esta función es para la actualización de cada subtotal
+    // de los productos, como para la actualización del valor
+    // total de la compra.
+    // Dichos cambios se generan al modificar la cantidad
+    // seleccionada para cada producto
+    actualizarTotal(e) {
         e.preventDefault();
         let id, cantidad, producto, productosLS;
         if (e.target.classList.contains('cantidad')) {
